# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wellezy prueba tecnica",
        "version": "0.1"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "description": "Autentica a un usuario con su correo electrónico y contraseña, devolviendo un token JWT.",
                "operationId": "5f90fad45b3bec06f69a5d179ce78a5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "3|VLypOerLIE2HlwDQYIoM8B1Osy7xAtzKHxCpSbWZk0HfLsXDb8fDRrGHocKvM1oQp6j"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error en la autenticación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se pudo iniciar sesión"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Contraseña o correo invalido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Algo salio mal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Algo salió mal"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Revoca todos los tokens de acceso del usuario autenticado, cerrando su sesión.",
                "operationId": "2baf9b9a1ed361caffdb6033ad366e94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LogoutRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sección cerrada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sección cerrada exitosamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al cerrar sesión",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Algo salió mal."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error detallado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "description": "Crea un nuevo usuario en el sistema y devuelve un token de acceso.",
                "operationId": "1716c95d7ee32950f11034c5cd88c6aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado y autenticado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "3|VLypOerLIE2HlwDQYIoM8B1Osy7xAtzKHxCpSbWZk0HfLsXDb8fDRrGHocKvM1oQp6j"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al registrar al usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al registrar al usuario."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error detallado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "camilo@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
            "LogoutRequest": {
                "required": [
                    "user_id"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Camilo Acacio"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "camilo@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "123456"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Ingresa tu token Bearer en el formato `Bearer {token}`",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}